Major rewrite of renderer interface in preparation for D3D12.
Renderer is now a fixed class, backend-dependant functionality now exists in RenderBackend and GPUDevice.

This will enable resource creation, as well as launching compute tasks from worker threads. Also cleans up a lot of the window messyness that was duplicated across render backends by consolidating it in the Renderer class.